Challenges/snakeCTF/meme-gallery/meme-app/requirements.txt
flask
python-mimeparse
gunicorn
minio
requests